requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-autoslug==1.9.9
django-countries==7.5.1
django-filter==23.2
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jwt==1.3.1
Markdown==3.4.4
phonenumbers==8.13.20
progressbar2==4.2.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-utils==3.7.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.1
typing_extensions==4.7.1
Unidecode==1.3.6
